+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Plist passed to `:lsp` now honors `:function` to set function other
+  than `harpoon-lsp-function` and `:dir-ignore-list` to set variable
+  other than `harpoon-lsp-dir-ignore-list`.
+
 ## [v0.2.0]
 
 ### Added